pyproject.toml
[project]
name = "job-application-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.21",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "apscheduler>=3.10,<4",
    "structlog>=24.4",
    "markdownify>=0.13",
    "markdown2>=2.4",
    "pypdf>=4.0",
    "python-docx>=1.0",
    "python-multipart>=0.0.9",
    "fastapi-users[sqlalchemy]>=13.0",
    "httpx-oauth>=0.13",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langgraph>=0.2",
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg[binary]>=3.1",
    "langsmith>=0.1",
    "sentry-sdk[fastapi]>=2.0",
    "weasyprint>=62.0",
    "greenlet>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "anyio>=4.0",
    "respx>=0.21",
    "testcontainers[postgres]>=4.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
